+++ src/ArduinoJson.hpp
@@ -4,6 +4,12 @@
 
 #pragma once
 
+#ifdef __clang__
+#pragma clang system_header
+#elif defined __GNUC__
+#pragma GCC system_header
+#endif
+
 #include "ArduinoJson/Namespace.hpp"
 
 #include "ArduinoJson/Array/ArrayRef.hpp"